Исправленное ДЗ по комментам
