+++ src/main/java/ru/javagrind/notify/Demo.java
@@ -1,13 +1,13 @@
-package Notify;
+package ru.javagrind.notify;
 
 public class Demo {
 
-    static String message = "Тестовое сообщение";
-    static Priority priority = Priority.NORMAL;
-    static String email = "qqq@qq.qq";
-    static String phoneNumber = "123123123";
-
     public static void main(String[] args) {
+
+        String message = "Тестовое сообщение";
+        Priority priority = Priority.NORMAL;
+        String email = "qqq@qq.qq";
+        String phoneNumber = "123123123";
 
         Notification notification = new Notification(priority, message);
         Notification emailNotification = new EmailNotification(message, email);
