+++ src/main/java/ru/javagrind/notify/EmailNotification.java
@@ -1,0 +1,26 @@
+package ru.javagrind.notify;
+
+public class EmailNotification extends Notification{
+    private String email;
+
+    public String getEmail() {
+        return email;
+    }
+
+    public EmailNotification(String message, String email) {
+        super(Priority.HIGH, message);
+        this.email = email;
+    }
+    @Override
+    public void send() {
+        System.out.println("Отправляем Email на ...");
+        super.send();
+
+//        System.out.print("id: " + getId() + ", ");
+//        System.out.println("Приоритет: " + getPriority());
+//        System.out.println("Отправляем Email на " + getEmail());
+//        System.out.println("Сообщение: " + getMessage());
+//        System.out.println();
+
+    }
+}
