+++ src/main/java/ru/javagrind/notify/Priority.java
@@ -1,4 +1,4 @@
-package Notify;
+package ru.javagrind.notify;
 
 public enum Priority {
     LOW, NORMAL, HIGH
